requirements.txt
numpy
pillow
opencv-python
ultralytics
matplotlib
torch
torchvision
